Move DataverseProvider to website/addons/dataverse/provider.py
